Some fixes on get one and on delete

+++ src/database/db.py
@@ -42,24 +42,28 @@ def get_one(model, identifier):
         data = session.query(model).get(identifier)
         session.commit()
 
-        return data, 200
+        if data:
+            return data, 200
+
+        return "Not Found!", 404
     except Exception as error:
         logger.error(error)
         session.rollback()
 
-        return str(error), 404
+        return str(error), 400
 
 def update(model, identifier, params):
     try:
         data = session.query(model).get(identifier)
         session.commit()
+        
         if data: 
             for param in params:
                 setattr(data, param, params[param])
             session.commit()
             return data, 200
-        else:
-            return "Not Found!", 404
+
+        return "Not Found!", 404
     except Exception as error:
         logger.error(error)
         session.rollback()
@@ -71,11 +75,15 @@ def delete(model, identifier):
 
     try:
         data = session.query(model).get(identifier)
-        session.delete(data)
-        session.commit()
-        return "Deleted successfully!", 204
+        
+        if data:
+            session.delete(data)
+            session.commit()
+            return "Deleted successfully!", 204
+
+        return "Not Found!", 404
     except Exception as error:
         logger.error(error)
         session.rollback()
 
-        return str(error), 404
+        return str(error), 400
